Updated Elimination System
Elimination System Updated.

Alive enemies that hasn't yet been damaged will also get an increase of health.
