+++ Devil's Punishment v2/Assets/Devil's Punishment Source/Data/Enemies/EnemyList.asset
@@ -14,4 +14,3 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   creatureHealthDatas:
   - {fileID: 11400000, guid: a97c831ea03b8e542820470b70c0b5fd, type: 2}
-  emptyList: []
